layer menu initially positioned left
