+++ src/components/MapLayerMenu/MapLayerMenuItem.tsx
@@ -22,6 +22,38 @@
  * SOFTWARE.
  */
 
-import LayerSelect from "./LayerSelect";
+import i18n from "@/i18n";
+import { LayerVisibilities } from "@/states/controlState";
+import SelectableMenuItem from "@/components/SelectableMenuItem";
 
-export default LayerSelect;
+interface LayerMenuItemProps {
+  layerId: keyof LayerVisibilities;
+  layerTitles: Record<keyof LayerVisibilities, string>;
+  layerSubtitles: Record<keyof LayerVisibilities, string>;
+  layerDisablements: Record<keyof LayerVisibilities, boolean>;
+  layerVisibilities: LayerVisibilities;
+  setLayerVisibility: (
+    layerId: keyof LayerVisibilities,
+    visible: boolean,
+  ) => void;
+}
+
+export default function MapLayerMenuItem({
+  layerId,
+  layerTitles,
+  layerSubtitles,
+  layerDisablements,
+  layerVisibilities,
+  setLayerVisibility,
+}: LayerMenuItemProps) {
+  const visible = !!layerVisibilities[layerId];
+  return (
+    <SelectableMenuItem
+      title={i18n.get(layerTitles[layerId])}
+      subtitle={layerSubtitles[layerId]}
+      disabled={layerDisablements[layerId]}
+      selected={visible}
+      onClick={() => setLayerVisibility(layerId, !visible)}
+    />
+  );
+}
